+++ src/commands.js
@@ -8,6 +8,8 @@ import {
   randomString,
   randomName,
   randomCity,
+  randomCountryCode,
+  randomCountryName,
   randomStreetAddress,
   randomPhoneNumber,
   randomEmail
@@ -23,6 +25,8 @@ export const extensionCommands = [
   { key: constants.CMD_KEY_RANDOM_GUID, callback: randomGuid },
   { key: constants.CMD_KEY_RANDOM_NAME, callback: randomName },
   { key: constants.CMD_KEY_RANDOM_CITY, callback: randomCity },
+  { key: constants.CMD_KEY_RANDOM_COUNTRY_CODE, callback: randomCountryCode },
+  { key: constants.CMD_KEY_RANDOM_COUNTRY_NAME, callback: randomCountryName },
   { key: constants.CMD_KEY_RANDOM_STREET_ADDRESS, callback: randomStreetAddress },
   { key: constants.CMD_KEY_RANDOM_PHONE_NUMBER, callback: randomPhoneNumber },
   { key: constants.CMD_KEY_RANDOM_EMAIL, callback: randomEmail }