+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to the "vscode-random" extension are documented in this file
 ## [1.1.0] - 2017-01-08
 ## Added
 - New command to allow random seed reset
+- New command to return a random country code
+- New command to return a random country name
 
 ## Changed
-- Refactoring to improve code simplicity and readability
+- Code improvements
 
 ## [1.0.2] - 2017-01-04
 ### Fixed